Remove all references to gogo-shell
The principal caller in liferay-theme-tasks was removed in:

https://github.com/liferay/liferay-js-themes-toolkit/pull/190

And we removed gogo-shell dead code from it in:

- https://github.com/liferay/liferay-js-themes-toolkit/pull/211 (v8)
- https://github.com/liferay/liferay-js-themes-toolkit/pull/212 (v9)

So let's remove it from here as well as a prelude to:

https://github.com/liferay/liferay-js-themes-toolkit/issues/193

We don't need to merge this one, but I wanted to document the changes
somewhere. This one builds on:

https://github.com/jbalsas/liferay-plugin-node-tasks/pull/2

Test plan: `npm run test` and feel the green.